fix controllers loading
